fix merging issues
