+++ ols/app/routers.py
@@ -2,7 +2,7 @@
 
 from fastapi import FastAPI
 
-from ols.app.endpoints import authorized, feedback, health, ols, streaming_ols
+from ols.app.endpoints import authorized, feedback, health, ols, streaming_ols, conversations
 from ols.app.metrics import metrics
 
 
@@ -15,6 +15,7 @@ def include_routers(app: FastAPI) -> None:
     app.include_router(ols.router, prefix="/v1")
     app.include_router(streaming_ols.router, prefix="/v1")
     app.include_router(feedback.router, prefix="/v1")
+    app.include_router(conversations.router)
     app.include_router(health.router)
     app.include_router(metrics.router)
     app.include_router(authorized.router)
